fix: change member variable into int

+++ 4_CPP_Module/09/ex01/RPN.hpp
@@ -16,5 +16,5 @@ private:
 	RPN(const RPN& copy);
 	RPN& operator=(const RPN& rhs);
 	
-	std::stack<unsigned long> mTokens;	
+	std::stack<int> mTokens;	
 };
